End point for start run simulation, moved ajax upload endpoint to sim
