+++ superset/templates/empower/lib.html
@@ -83,7 +83,7 @@
                     });
                     $.ajax({
                         type: "POST",
-                        url: "/superset/upload_assumption_ajax",
+                        url: "/simulationmodelview/upload_assumption_ajax",
                         xhr: function () {
                             var myXhr = $.ajaxSettings.xhr();
                             return myXhr;
@@ -95,10 +95,9 @@
                             new_option.value = data['detail']['id']
                             new_option.innerHTML = data['detail']['name']
                             document.getElementById('assumption').appendChild(new_option)
-                            console.log(document.getElementById('assumption'))
                             button.innerHTML = 'Upload'
                         },
-                        complete: function(x, s){
+                        error: function(x, s){
                             if (s =='timeout'){
                                 alert('Timeout, please try again later.')
                             }
@@ -125,6 +124,37 @@
 
 
 {% macro render_form_controls() %}
+    <div class="well well-sm">
+        <button id="submit_button" type="submit" class="btn  btn-sm btn-primary">{{_("Save")}}
+            <i class="fa fa-save"></i></button>
+        {{ lib.lnk_back() }}
+        <script>
+            // $('#submit_button').click( function(e) {
+            //     e.preventDefault();
+            //     let assumption_choice;
+            //
+            //     // list = []
+            //     let assumption = document.getElementById('assumption')
+            //     console.log(assumption)
+            //     // for(var i = 0, row; row = table.rows[i]; i++) {
+            //     //     if (row.id != ''){
+            //     //         list.push(row.id);
+            //     //     }
+            //     // }
+            //     // form = document.getElementById('model_form');
+            //     // var sites_list = document.createElement("input");
+            //     // sites_list.type = 'hidden';
+            //     // sites_list.name = 'sites';
+            //     // sites_list.id = 'sites';
+            //     // sites_list.value = list;
+            //     // form.appendChild(sites_list);
+            //     form.submit()
+            // })
+        </script>
+    </div>
+{% endmacro %}
+
+{% macro render_save_run() %}
     <div class="well well-sm">
         <button id="submit_button" type="submit" class="btn  btn-sm btn-primary">{{_("Save")}}
             <i class="fa fa-save"></i></button>
