+++ superset/views/simulation/views.py
@@ -20,7 +20,7 @@ import traceback
 import logging
 import os
 import tempfile
-from flask import flash, redirect, g, request, abort, url_for
+from flask import flash, redirect, g, request, abort, url_for, jsonify
 from flask_appbuilder.widgets import ListWidget, FormWidget, ShowWidget
 from flask_appbuilder import expose, has_access, SimpleFormView
 from flask_appbuilder.urltools import get_filter_args, get_order_args, get_page_args, get_page_size_args
@@ -439,7 +439,10 @@ class SimulationModelView(
 
     route_base = "/simulationmodelview"
     datamodel = SQLAInterface(Simulation)
-    include_route_methods = RouteMethod.CRUD_SET
+    include_route_methods = RouteMethod.CRUD_SET | {
+        'upload_assumption_ajax',
+        'start_run',
+    }
 
     list_columns = ['run_id', 'name','assumption', 'project', 'status']
     # add_columns = ['name','run_id','description','assumption_choice1','assumption_choice2','generation_model',
@@ -509,7 +512,6 @@ class SimulationModelView(
         )
         return widgets
 
-
     @simulation_logger.log_simulation(action_name='create simulation')
     def add_item(self, form, exclude_cols):
         self._fill_form_exclude_cols(exclude_cols, form)
@@ -575,6 +577,35 @@ class SimulationModelView(
             form.report_type.data = simulation.report_type
         return form
 
+    @simulation_logger.log_simulation(action_name='upload assumption')
+    @expose('/upload_assumption_ajax', methods=['POST'])
+    def upload_assumption_ajax(self):
+        file = request.files['file']
+        try:
+            g.action_object = file.filename
+            g.action_object_type = 'Assumption'
+            path = os.path.join(app.config['UPLOAD_FOLDER'], file.filename)
+            file.save(path)
+            file_name = file.filename.split('.')[0]
+            id, name = upload_assumption_and_trigger_process(path, file_name)
+            message = 'Uploaded success. You can now choose the uploaded assumption in the list. The assumption ' + \
+                      'processing is running on backend and will be ready soon.'
+            detail = {'name': name,
+                      'id': id}
+            g.result = 'Upload success, processing'
+            g.detail = None
+        except Exception as e:
+            message = 'Failed. {}'.format(repr(e))
+            detail = repr(e)
+            g.result = 'Upload failed'
+            g.detail = detail
+        finally:
+            # os.remove(os.path.join(app.config['UPLOAD_FOLDER'], file.filename))
+            return jsonify({
+                'message': message,
+                'detail': detail
+            })
+
     def prefill_hidden_field(self, form):
         flt_dic = self.get_filter_args()
         # print(form)
@@ -594,6 +625,27 @@ class SimulationModelView(
             if sim_match:
                 result['sim'] = request.args.get(arg)
         return result
+
+    @simulation_logger.log_simulation(action_name='start run')
+    @expose('/start_run/<id>/<run_type>/')
+    def start_run(self,id,run_type):
+        simulation = db.session.query(Simulation).filter_by(id=id).one_or_none()
+        message = detail = None
+        if not simulation:
+            message = 'No simulation found for this run id, please refresh the page and try again.'
+            g.result = 'Run failed'
+            g.detail = message
+            detail = None
+        if run_type == 'test':
+            pass
+        else:
+            pass
+        return jsonify({
+            'message': message,
+            'detail': detail
+        })
+
+
 
 
 class ProjectModelView(EmpowerModelView):
